requirements.txt
flask
peewee
flask-compress
ruamel-yaml
authlib
requests
flask-wtf
